+++ apps/orchestrator/src/components/modals/ClaudeHistoryModal.tsx
@@ -1,12 +1,13 @@
 import { useEffect, useState } from 'react'
 
 import { preparePtyRuntimeLaunch } from '../../lib/agentRuntimeAdapter'
-import { intlLocale, useT, type Locale, type TFunction } from '../../lib/i18n'
-import { listClaudeSessions, restartPty, type ClaudeSessionMeta } from '../../lib/tauri'
+import { intlLocale, type Locale, type TFunction,useT } from '../../lib/i18n'
+import { providerSpawnFlags } from '../../lib/providers/envForCli'
+import { type ClaudeSessionMeta,listClaudeSessions, restartPty } from '../../lib/tauri'
 import { agentCliCommand, type AgentType } from '../../lib/types'
 import { useProjectsStore } from '../../stores/projectsStore'
-import { Modal } from './Modal'
 import styles from './ClaudeHistoryModal.module.css'
+import { Modal } from './Modal'
 
 type Props = {
   open: boolean
@@ -100,6 +101,7 @@ export function ClaudeHistoryModal({
         cwd,
         extraArgs: preparedRuntime.args,
         env: preparedRuntime.env,
+        ...providerSpawnFlags(agentType, useProjectsStore.getState().preferences),
       })
       window.dispatchEvent(new CustomEvent('flashwork:terminal-resize-request', { detail: { ptyId } }))